+++ packages/eas-cli/src/worker/fragments/WorkerDeploymentAlias.ts
@@ -1,0 +1,9 @@
+import gql from 'graphql-tag';
+
+export const WorkerDeploymentAliasFragmentNode = gql`
+  fragment WorkerDeploymentAliasFragment on WorkerDeploymentAlias {
+    id
+    aliasName
+    url
+  }
+`;